Split time machine test suite into separate files
